fix: clip website window away from sidebar
